refactor(verifiability): delete dead showcase runner layer
