+++ docs/topology-status.md
@@ -28,10 +28,10 @@ Generated by `devtools render topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 697
+- **Stable** (no move scoped): 693
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 16
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 838
-- **Realized polylogue/**/*.py**: 838 files declared
+- **Total declared**: 834
+- **Realized polylogue/**/*.py**: 834 files declared
 <!-- END GENERATED: topology-status -->
